\name{retour_op}
\alias{retour_op}
\title{retour op}
\description{This function is used by the Fpop function to recover the best
segment ends from 1:n from the C output.}
\usage{retour_op(path)}
\arguments{
  \item{path}{the path vector of the "colibri_op_R_c C" function}
}

\value{a vector with the best segment ends.}

\author{Guillem Rigaill, Toby Dylan Hocking}





